+++ packages/api/src/platforms/vtex/loaders/salesChannel.ts
@@ -1,15 +1,13 @@
-import DataLoader from "dataloader";
+import DataLoader from 'dataloader'
 
-import { Options } from "..";
-import { Clients } from "../clients";
+import type { Options } from '..'
+import type { Clients } from '../clients'
 
-import type { SalesChannel } from "./../clients/commerce/types/SalesChannel";
+import type { SalesChannel } from './../clients/commerce/types/SalesChannel'
 
 export const getSalesChannelLoader = (_: Options, clients: Clients) => {
   const loader = async (channels: readonly string[]) =>
-    Promise.all(
-      channels.map((sc) => clients.commerce.catalog.salesChannel(sc)),
-    );
+    Promise.all(channels.map((sc) => clients.commerce.catalog.salesChannel(sc)))
 
-  return new DataLoader<string, SalesChannel>(loader);
-};
+  return new DataLoader<string, SalesChannel>(loader)
+}