+++ packages/cli/src/utils/generate.ts
@@ -57,7 +57,7 @@ function filterAndCopyPackageJson(basePath: string) {
   const corePackageJsonPath = path.join(coreDir, 'package.json')
 
   const corePackageJsonFile = readFileSync(corePackageJsonPath, 'utf8')
-  let { exports: _, ...filteredFileContent } = JSON.parse(corePackageJsonFile)
+  const { exports: _, ...filteredFileContent } = JSON.parse(corePackageJsonFile)
 
   filteredFileContent.name = 'dot-faststore'
 
@@ -396,7 +396,8 @@ function updateNextConfig(basePath: string) {
 
 // returns new (discovery.config.js) or legacy (faststore.config.js) store config file
 function getCurrentUserStoreConfigFile(basePath: string) {
-  const { userStoreConfigFile, userLegacyStoreConfigFile } = withBasePath(basePath)
+  const { userStoreConfigFile, userLegacyStoreConfigFile } =
+    withBasePath(basePath)
 
   if (existsSync(userStoreConfigFile)) {
     return userStoreConfigFile
@@ -412,7 +413,7 @@ function getCurrentUserStoreConfigFile(basePath: string) {
 function validateAndInstallMissingDependencies(basePath: string) {
   const { userDir } = withBasePath(basePath)
 
-  const currentUserStoreConfigFile = getCurrentUserStoreConfigFile(basePath) 
+  const currentUserStoreConfigFile = getCurrentUserStoreConfigFile(basePath)
 
   if (!currentUserStoreConfigFile) {
     return
@@ -492,11 +493,11 @@ function enableRedirectsMiddleware(basePath: string) {
 function enableSearchSSR(basePath: string) {
   const storeConfigPath = getCurrentUserStoreConfigFile(basePath)
 
-  if(!storeConfigPath) { 
-    return 
+  if (!storeConfigPath) {
+    return
   }
   const storeConfig = require(storeConfigPath)
-  if(!storeConfig.experimental.enableSearchSSR) { 
+  if (!storeConfig.experimental.enableSearchSSR) {
     return
   }
 
@@ -504,7 +505,10 @@ function enableSearchSSR(basePath: string) {
   const searchPagePath = path.join(tmpDir, 'src', 'pages', 's.tsx')
   const searchPageData = String(readFileSync(searchPagePath))
 
-  const searchPageWithSSR = searchPageData.replaceAll('getStaticProps', 'getServerSideProps')
+  const searchPageWithSSR = searchPageData.replaceAll(
+    'getStaticProps',
+    'getServerSideProps'
+  )
 
   writeFileSync(searchPagePath, searchPageWithSSR)
 }