+++ packages/core/discovery.config.default.js
@@ -6,11 +6,13 @@ module.exports = {
     author: 'Store Framework',
     plp: {
       titleTemplate: '%s | FastStore PLP',
-      descriptionTemplate: '%s products on FastStore Product Listing Page'
+      descriptionTemplate: '%s products on FastStore Product Listing Page',
     },
     search: {
       titleTemplate: '%s: Search results title',
       descriptionTemplate: '%s: Search results description',
+      noIndex: true,
+      noFollow: true,
     },
   },